% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdown-package.R
\name{github_html_document}
\alias{github_html_document}
\title{Create a Github styled HTML document}
\usage{
github_html_document(...)
}
\arguments{
\item{...}{parameters passed onto \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
\code{\link[rmarkdown:github_document]{rmarkdown::github_document()}} can create a temporary HTML preview of an
RMarkdown file in order to show what a markdown document would look like on
Github. This function is a wrapper \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}} that uses the
Github CSS style used in the HTML previews created by \code{github_document()}.
}
